added current weather details

+++ app/src/main/java/com/example/weather_app/remote/model/CoordinatesResponse.kt
@@ -1,6 +1,6 @@
 package com.example.weather_app.remote.model
 
 data class CoordinatesResponse(
-    val lon: Float,
-    val lat: Float
+    val lon: Double,
+    val lat: Double
 )
